Migrate canonical-only test
