+++ cypress/integration/pages/articles/testsForCanonicalOnly.js
@@ -1,5 +1,6 @@
 import envConfig from '../../../support/config/envs';
 import config from '../../../support/config/services';
+import { getBlockData } from './helpers';
 
 // TODO: Remove after https://github.com/bbc/simorgh/issues/2959
 const serviceHasCaption = service => service === 'news';
@@ -68,6 +69,24 @@ export const testsThatFollowSmokeTestConfigForCanonicalOnly = ({
           });
       });
     }
+
+    it('should render a placeholder image within a media block', () => {
+      cy.request(`${config[service].pageTypes.articles.path}.json`).then(
+        ({ body }) => {
+          // `video` blocks can also contain audio, so this test
+          // checks if both media types have a placeholder image.
+          const media = getBlockData('video', body);
+          if (media) {
+            cy.get('div[class^="StyledVideoContainer"]').within(() => {
+              cy.get('div[class^="StyledPlaceholder"] > img')
+                .should('be.visible')
+                .should('have.attr', 'src')
+                .should('not.be.empty');
+            });
+          }
+        },
+      );
+    });
   });
 
 // For testing low priority things e.g. cosmetic differences, and a safe place to put slow tests.
